requirements.txt
log-to-kafka
multi-thread-closing